フォロー機能追加

+++ database/migrations/2021_08_18_213438_create_microposts_table.php
@@ -15,7 +15,7 @@ class CreateMicropostsTable extends Migration
     {
         Schema::create('microposts', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('user_id')->unsigned()->index();
+            $table->unsignedBigInteger('user_id')->unsigned()->index();
             $table->string('content');
             $table->timestamps();
